+++ crates/dioxus-class/src/build.rs
@@ -1,4 +1,5 @@
 use crate::prelude::*;
+use std::fs;
 use std::fs::File;
 use std::io::LineWriter;
 use std::io::Write;
@@ -41,4 +42,35 @@ pub fn generate_classes<P: AsRef<Path> + Clone>(
         class_count, size, path_str
     );
     Ok(())
+}
+
+pub fn check_vec_ending(content: &str) -> bool {
+    for ch in content.chars().rev() {
+        if ch == ']' {
+            return true;
+        } else if vec!['\n', '\r', ' ', '\t'].contains(&ch) {
+            continue;
+        } else {
+            return false;
+        }
+    }
+    false
+}
+
+pub fn check_classes<P: AsRef<Path> + Clone>(
+    path: P,
+) -> Result<(), Box<dyn std::error::Error>> {
+    let content = if path.as_ref().exists() {
+        let lines = fs::read_to_string(path.clone())?;
+        if check_vec_ending(&lines) {
+            return Ok(())
+        }
+        println!("cargo:warning=dioxus-class::build::check_classes: append vec ending");
+        lines + "]\n"
+    } else {
+        println!("cargo:warning=dioxus-class::build::check_classes: create empty file");
+
+        "vec![]\n".to_owned()
+    };
+    write_file(path, &content)
 }