+++ demo/emoji-browser/css/classes.rs
@@ -1,0 +1,82 @@
+vec![
+
+/* #257 bytes(434..539)
+class!(card card_compact w_64 h_64 bg_base_300 shadow_xl text_center hover(bg_base_200) hover(scale_105))
+ */
+Class ::
+from(vec!
+[String :: from(card), String :: from(card_compact), String :: from(w_64),
+String :: from(h_64), String :: from(bg_base_300), String :: from(shadow_xl),
+String :: from(text_center), String :: from(hover(bg_base_200)), String ::
+from(hover(scale_105))]),
+
+/* #263 bytes(582..604)
+class!(text_8xl py_10)
+ */
+Class :: from(vec! [String :: from(text_8xl), String :: from(py_10)]),
+
+/* #268 bytes(689..731)
+class!(card_body text_center items_center)
+ */
+Class ::
+from(vec!
+[String :: from(card_body), String :: from(text_center), String ::
+from(items_center)]),
+
+/* #273 bytes(782..826)
+class!(card_title text_sm text_base_content)
+ */
+Class ::
+from(vec!
+[String :: from(card_title), String :: from(text_sm), String ::
+from(text_base_content)]),
+
+/* #293 bytes(1067..1124)
+class!(flex flex_row flex_wrap justify_center gap_4 mx_2)
+ */
+Class ::
+from(vec!
+[String :: from(flex), String :: from(flex_row), String :: from(flex_wrap),
+String :: from(justify_center), String :: from(gap_4), String :: from(mx_2)]),
+
+/* #316 bytes(1376..1446)
+class!(w_80 h_12 text_2xl m_4 px_2 input input_primary input_bordered)
+ */
+Class ::
+from(vec!
+[String :: from(w_80), String :: from(h_12), String :: from(text_2xl), String
+:: from(m_4), String :: from(px_2), String :: from(input), String ::
+from(input_primary), String :: from(input_bordered)]),
+
+/* #324 bytes(2028..2057)
+class!(bottom_0 h_full mt_20)
+ */
+Class ::
+from(vec!
+[String :: from(bottom_0), String :: from(h_full), String :: from(mt_20)]),
+
+/* #329 bytes(2101..2115)
+class!(hidden)
+ */
+Class :: from(vec! [String :: from(hidden)]),
+
+/* #334 bytes(2186..2200)
+class!(hidden)
+ */
+Class :: from(vec! [String :: from(hidden)]),
+
+/* #362 bytes(2269..2294)
+class!(w_screen h_screen)
+ */
+Class :: from(vec! [String :: from(w_screen), String :: from(h_screen)]),
+
+/* #367 bytes(2337..2400)
+class!(fixed top_0 w_full z_10 bg_base_100 flex justify_center)
+ */
+Class ::
+from(vec!
+[String :: from(fixed), String :: from(top_0), String :: from(w_full), String
+:: from(z_10), String :: from(bg_base_100), String :: from(flex), String ::
+from(justify_center)]),
+
+]