+++ crates/dioxus-class-macro/CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.7.0] - 2024-04-09
+
+- Support DIOXUS_CLASS_BUILD_PATH env to control where to write the classes.rs file
+- Tweak logging for classes build process
+
 ## [0.5.0] - 2024-03-28
 
 ### Added