+++ crates/dioxus-class/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.7.0] - 2024-04-09
+
+- Add build::check_classes to create empty file and adding vec ending
+
 ## [0.6.0] - 2024-04-01
 
 ### Fixed